+++ modules/spring-core/src/main/kotlin/com/esprito/spring/core/externalsystem/process/UnlinkedProjectAware.kt
@@ -5,13 +5,17 @@ import com.esprito.spring.core.externalsystem.setting.NativeSettings
 import com.esprito.spring.core.externalsystem.utils.Constants
 import com.esprito.spring.core.externalsystem.utils.NativeBootUtils
 import com.esprito.spring.core.externalsystem.utils.NativeBootUtils.getVirtualFile
+import com.esprito.spring.core.tracker.ModificationTrackerManager
+import com.intellij.codeInsight.daemon.DaemonCodeAnalyzer
 import com.intellij.openapi.Disposable
+import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.externalSystem.autolink.ExternalSystemProjectLinkListener
 import com.intellij.openapi.externalSystem.autolink.ExternalSystemUnlinkedProjectAware
 import com.intellij.openapi.externalSystem.settings.ExternalSystemSettingsListener
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.projectImport.ProjectOpenProcessor
+import com.intellij.psi.PsiManager
 
 
 class UnlinkedProjectAware : ExternalSystemUnlinkedProjectAware {
@@ -31,7 +35,7 @@ class UnlinkedProjectAware : ExternalSystemUnlinkedProjectAware {
         project: Project, listener: ExternalSystemProjectLinkListener, parentDisposable: Disposable
     ) {
         val nativeSettings = project.getService(NativeSettings::class.java)
-        nativeSettings.subscribe(UnlinkSettingsListener(listener), parentDisposable)
+        nativeSettings.subscribe(UnlinkSettingsListener(project, listener), parentDisposable)
     }
 
     override fun linkAndLoadProject(project: Project, externalProjectPath: String) {
@@ -41,13 +45,18 @@ class UnlinkedProjectAware : ExternalSystemUnlinkedProjectAware {
     }
 }
 
-internal class UnlinkSettingsListener(val listener: ExternalSystemProjectLinkListener) :
-    ExternalSystemSettingsListener<NativeProjectSettings?> {
-    override fun onProjectsLinked(settings: MutableCollection<NativeProjectSettings?>) {
-        settings.filterNotNull().forEach { listener.onProjectLinked(it.externalProjectPath) }
+internal class UnlinkSettingsListener(val project: Project, val listener: ExternalSystemProjectLinkListener) :
+    ExternalSystemSettingsListener<NativeProjectSettings> {
+    override fun onProjectsLinked(settings: MutableCollection<NativeProjectSettings>) {
+        settings.forEach { listener.onProjectLinked(it.externalProjectPath) }
     }
 
     override fun onProjectsUnlinked(linkedProjectPaths: Set<String>) {
         linkedProjectPaths.forEach { listener.onProjectUnlinked(it) }
+        ApplicationManager.getApplication().invokeLater {
+            ModificationTrackerManager.getInstance(project).invalidateAll()
+            PsiManager.getInstance(project).dropPsiCaches()
+            DaemonCodeAnalyzer.getInstance(project).restart()
+        }
     }
 }